requirements.txt
# Core LLM / Agent framework
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# HTTP & data
httpx>=0.27.0
curl_cffi>=0.7.0
pandas>=2.2.0

# Environment
python-dotenv>=1.0.0

# (Optional) pretty console output
rich>=13.0.0

# Web UI
streamlit>=1.30.0
